+++ CHANGELOG.md
@@ -1,3 +1,34 @@
+### 1.2.6
+
+## Enhancements
+
+- Updated the native iOS agent to version 7.7.2.
+
+### 1.2.5
+
+## Enhancements
+
+- Updated the native Android agent to version 7.7.5.
+
+### 1.2.4
+
+## Enhancements
+
+- Updated the native Android agent to version 7.7.4.
+
+### 1.2.3
+
+## Enhancements
+
+- Updated the native Android agent to version 7.7.2.
+- Updated the native iOS agent to version 7.7.1.
+
+### 1.2.2
+
+## Enhancements
+
+- Updated the native iOS agent to version 7.7.0. 
+
 ### 1.2.1
 
 ## Enhancements